delete confirmation

+++ production.php
@@ -340,7 +340,8 @@ if(!empty(($_FILES["file"]["name"]))){
                                 <br>
                                 <div class="row mt-3">
                                     <div class="col-4 text-start">
-                                        <a href="production.php?delete=<?php echo $id; ?>" class="text-uppercase  fs-6 btn btn-danger btn-sm">Delete</a>
+                                        <a href="production.php?delete=<?php echo $id; ?>" class="text-uppercase  fs-6 btn btn-danger btn-sm"
+                                          onclick="confirmMessage()" >Delete</a>
                                     </div>
 
                                     <div class="col-4 text-center">
@@ -374,7 +375,6 @@ if(!empty(($_FILES["file"]["name"]))){
                                         }
                                         ?>
 
-
                                     </div>
 
                                     <div class="col-4 text-end">
@@ -383,21 +383,26 @@ if(!empty(($_FILES["file"]["name"]))){
 
                                 </div>
 
+                                <script>
+                                    function confirmMessage(){
+                                        if(confirm("You Sure? You Want To Delete")){
+                                            <?php
+                                            //delete the entries
+                                            if(isset($_GET['delete'])){
+                                            $idGet=$_GET['delete'];
+                                            $queryGet="DELETE FROM production WHERE id= '$idGet'";
+                                            $resultGet=mysqli_query($connection,$queryGet);
+                                            if(!$resultGet){
+                                                die("not deleted" . mysqli_error($connection));
+                                            }?>
+                                                window.location.href="production.php";
                                 <?php
-                                //delete the entries
-                                if(isset($_GET['delete'])){
-                                    $idGet=$_GET['delete'];
-                                    $queryGet="DELETE FROM production WHERE id= '$idGet'";
-                                    $resultGet=mysqli_query($connection,$queryGet);
-                                    if(!$resultGet){
-                                        die("not deleted" . mysqli_error($connection));
-                                    }?>
-                                    <script>
-                                        window.location.href="production.php";
-                                    </script>
-                                    <?php
                                 }
                                 ?>
+                                        }
+                                    }
+                                </script>
+
                             </div>
                         </div>
                     </div>
